+++ CHANGELOG.md
@@ -602,6 +602,8 @@
 - Update Rails configuration files with v6.1 settings
 - Add tests to guard against missing Activity field translations
 - Provide a breakdown of spend, refund and net values as a report CSV download
+- Only include transferred and direct budget in budget calculations
+- Show an activity's total forecasted spend to date against an activity
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-45...HEAD
 [release-45]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-44...release-45